+++ AzeriteUI/modules/unitframe_targetframe.lua
@@ -567,6 +567,7 @@ local Style = function(self, unit, id, ...)
 	-----------------------------------------------------------	
 	local threats = { IsShown = Threat_IsShown, Show = Threat_Show, Hide = Threat_Hide }
 	threats.feedbackUnit = "player"
+	threats.hideSolo = true
 
 	local threatHealth = backdrop:CreateTexture()
 	threatHealth:SetDrawLayer("BACKGROUND", -2)